+++ dca/tl_wem_portfolio_item.php
@@ -165,6 +165,28 @@ $GLOBALS['TL_DCA']['tl_wem_portfolio_item'] = array
 			),
 			'sql'                     => "varchar(128) COLLATE utf8_bin NOT NULL default ''"
 		),
+		'customer' => array
+		(
+			'label'                   => &$GLOBALS['TL_LANG']['tl_wem_portfolio_item']['customer'],
+			'exclude'                 => true,
+			'inputType'               => 'select',
+			'foreignKey'              => 'tl_wem_portfolio_customer.title',
+			'eval'                    => array('doNotCopy'=>true, 'chosen'=>true, 'includeBlankOption'=>true, 'tl_class'=>'w50'),
+			'sql'                     => "int(10) unsigned NOT NULL default '0'",
+			'relation'                => array('type'=>'belongsTo', 'load'=>'lazy')
+		),
+		'date' => array
+		(
+			'label'                   => &$GLOBALS['TL_LANG']['tl_wem_portfolio_item']['date'],
+			'default'                 => time(),
+			'exclude'                 => true,
+			'filter'                  => true,
+			'sorting'                 => true,
+			'flag'                    => 8,
+			'inputType'               => 'text',
+			'eval'                    => array('rgxp'=>'date', 'doNotCopy'=>true, 'datepicker'=>true, 'tl_class'=>'w50 wizard'),
+			'sql'                     => "int(10) unsigned NOT NULL default '0'"
+		),
 		'picture' => array
 		(
 			'label'                   => &$GLOBALS['TL_LANG']['tl_wem_portfolio_item']['picture'],