+++ CHANGELOG.md
@@ -2,5 +2,12 @@
 Extension "Portfolio" for Contao Open Source CMS
 ========
 
-V0.1 - 2017-11-03
+0.1 - 2017-11-03
 - Init git repository
+
+0.1.1 - 2017-11-06
+- Composer tests
+
+0.1.2 - 2017-11-08
+- Add severals fields to extends configuration
+- Start the frontend modules